+++ .next/server/server-reference-manifest.js
@@ -1,1 +1,1 @@
-self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"4efuEw5EzYsjO2ruf7Pl2IrYnC1JVUbQ8s5YY0QzWLQ=\"\n}"
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"rASg3RXziogYf3O5WHcqZROrVxyljpscCnK8afIjiBA=\"\n}"